+++ lib/column/json.go
@@ -576,15 +576,32 @@ func (c *JSON) encodeStringData(buffer *proto.Buffer) {
 	c.jsonStrings.Encode(buffer)
 }
 
-func (c *JSON) Encode(buffer *proto.Buffer) {
+func (c *JSON) WriteStatePrefix(buffer *proto.Buffer) error {
 	switch c.serializationVersion {
 	case JSONObjectSerializationVersion:
 		buffer.PutUInt64(JSONObjectSerializationVersion)
 		c.encodeObjectHeader(buffer)
+
+		return nil
+	case JSONStringSerializationVersion:
+		buffer.PutUInt64(JSONStringSerializationVersion)
+
+		return nil
+	default:
+		// If the column is an array, it can be empty but still require a prefix.
+		// Use string encoding since it's smaller
+		buffer.PutUInt64(JSONStringSerializationVersion)
+
+		return nil
+	}
+}
+
+func (c *JSON) Encode(buffer *proto.Buffer) {
+	switch c.serializationVersion {
+	case JSONObjectSerializationVersion:
 		c.encodeObjectData(buffer)
 		return
 	case JSONStringSerializationVersion:
-		buffer.PutUInt64(JSONStringSerializationVersion)
 		c.encodeStringData(buffer)
 		return
 	}
@@ -686,9 +703,7 @@ func (c *JSON) decodeStringData(reader *proto.Reader, rows int) error {
 	return c.jsonStrings.Decode(reader, rows)
 }
 
-func (c *JSON) Decode(reader *proto.Reader, rows int) error {
-	c.rows = rows
-
+func (c *JSON) ReadStatePrefix(reader *proto.Reader) error {
 	jsonSerializationVersion, err := reader.UInt64()
 	if err != nil {
 		return fmt.Errorf("failed to read json serialization version: %w", err)
@@ -703,20 +718,34 @@ func (c *JSON) Decode(reader *proto.Reader, rows int) error {
 			return fmt.Errorf("failed to decode json object header: %w", err)
 		}
 
-		err = c.decodeObjectData(reader, rows)
+		return nil
+	case JSONStringSerializationVersion:
+		return nil
+	default:
+		return fmt.Errorf("unsupported JSON serialization version for prefix decode: %d", jsonSerializationVersion)
+	}
+}
+
+func (c *JSON) Decode(reader *proto.Reader, rows int) error {
+	c.rows = rows
+
+	switch c.serializationVersion {
+	case JSONObjectSerializationVersion:
+		err := c.decodeObjectData(reader, rows)
 		if err != nil {
 			return fmt.Errorf("failed to decode json object data: %w", err)
 		}
 
 		return nil
 	case JSONStringSerializationVersion:
-		err = c.decodeStringData(reader, rows)
+		err := c.decodeStringData(reader, rows)
 		if err != nil {
 			return fmt.Errorf("failed to decode json string data: %w", err)
 		}
+
 		return nil
 	default:
-		return fmt.Errorf("unsupported JSON serialization version for decode: %d", jsonSerializationVersion)
+		return fmt.Errorf("unsupported JSON serialization version for decode: %d", c.serializationVersion)
 	}
 }